remove Features.java
Starting from this commit, we will try to minimize the use of intermediate Java classes and make the most of what the NDK provides, or use JNI when necessary, but by directly calling the Android API rather than other intermediate Java classes.

This commit is therefore a redesign of the feature system. It revises the `gradle.properties` file to now only contain `display.xxx` parameters for configuring the display, and `requirements.xxx`, which will simply indicate in the `AndroidManifest.xml` which features the devices must have to run the application.

This also involved a redesign of the system managing sensors, which is now much simpler to use and written entirely in native code.

Additionally, an improvement was made to vibration management, which now includes functions to control the vibration time in milliseconds as well as functions taking an intensity parameter. This system is also now written in native code, using JNI but without using any intermediate classes.